feat: adding arrow sign for Competition landing Carousel
